requirements.txt
virtualenv
django
djangorestframework
PyMySQL
celery
django-celery-results
django-celery-beat
